deleted base controller, fixed module controller
