+++ app/controllers/admin/dashboard_controller.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
 module Admin
-  class DashboardController < BaseController
-    load_and_authorize_resource class: false
-  end
+  class DashboardController < ModuleController; end
 end
