+++ app/models/ability.rb
@@ -34,7 +34,7 @@ class Ability
     # See the wiki for details:
     # https://github.com/CanCanCommunity/cancancan/wiki/Defining-Abilities
 
-    case account.role
+    case account&.role
       when 'admin'
         can :manage, :all
       when 'compliance'
